Max reader algo (#1352)
For date and id slicer/reader
two new feature flags for experimentation:

- `total_optimization`: defaults to false, if set to true, will will
change the track_total_hits parameter inside the query to not find the
exact total count for a given query. This is useful for dealing with
slices and reads of significant data density. Setting this to true, you
do not gain much benefit at all with the `recurse_optimization` flag so
its advised not to use them both together.
- `recurse_optimization`: defaults to false, if set to true, it attempts
to lower the amount of times it has to recount a query to make a slice
by make smarter chunk sizes based off the total count instead of its
simple splitting logic, only really useful if `total_optimization` is
set to false

For spaces slicer/reader
The two new feature flags for experimentation are the same except that
`total_optimization` defaults to true as per the current usage of this
does so in spaces